renamed processor.py to get_resonance_signal; moved identify_1234/5230 from utils to get_resonance_signal
